Move function that returns query options next to hook
